refactor(rtmp): deduplicate RTMP chunk parser implementations
