+++ src/projects/modules/rtmp/chunk/rtmp_chunk_parser.cpp
@@ -8,7 +8,10 @@
 //==============================================================================
 #include "rtmp_chunk_parser.h"
 
+// `rtmp_private.h` should be included before `rtmp_chunk_parser_helper.h` because it contains `OV_LOG_TAG`.
 #include "../rtmp_private.h"
+// ====
+#include "../rtmp_chunk_parser_helper.h"
 
 namespace
 {
@@ -287,84 +290,6 @@ bool RtmpChunkParser::IsContinuationChunk(const uint32_t chunk_stream_id) const
 	return false;
 }
 
-std::optional<uint32_t> RtmpChunkParser::ParseTimestampField(
-	const uint32_t stream_id,
-	ov::ByteStream &stream,
-	RtmpChunkHeader *chunk_header,
-	const uint32_t encoded_value)
-{
-	if (encoded_value != EXTENDED_TIMESTAMP_INDICATOR)
-	{
-		chunk_header->is_extended_timestamp = false;
-		chunk_header->extended_timestamp	= 0U;
-
-		return encoded_value;
-	}
-
-	if (stream.IsRemained(EXTENDED_TIMESTAMP_SIZE) == false)
-	{
-		logap("Need more data to parse extended timestamp field: %d bytes (current: %zu)", EXTENDED_TIMESTAMP_SIZE, stream.Remained());
-		return std::nullopt;
-	}
-
-	logat("Extended timestamp is present for stream id: %u", stream_id);
-
-	const auto extended_timestamp		= stream.ReadBE32();
-
-	chunk_header->is_extended_timestamp = true;
-	chunk_header->extended_timestamp	= extended_timestamp;
-	chunk_header->message_header_length += EXTENDED_TIMESTAMP_SIZE;
-
-	return extended_timestamp;
-}
-
-std::optional<int64_t> RtmpChunkParser::ResolveTimestampDelta(
-	const uint32_t stream_id,
-	const int64_t preceding_timestamp,
-	const uint32_t timestamp_delta,
-	const bool is_extended_timestamp) const
-{
-	if (is_extended_timestamp && (timestamp_delta >= SIGNED_EXTENDED_TIMESTAMP_DELTA_SIGN_BIT))
-	{
-		// The sign bit is set in a 32-bit extended delta, so this could be a
-		// broken sender that wrote a signed int32_t onto the wire.
-		const auto signed_timestamp_delta = static_cast<int32_t>(timestamp_delta);
-		const auto backward_delta		  = -1 * static_cast<int64_t>(signed_timestamp_delta);
-
-		if ((signed_timestamp_delta < 0) && (backward_delta <= MAX_NEGATIVE_EXTENDED_TIMESTAMP_DELTA_MS))
-		{
-			// Accept only small backward jumps on the compatibility path. Large
-			// values are more likely to be legitimate unsigned deltas.
-			const auto resolved_timestamp = preceding_timestamp + signed_timestamp_delta;
-
-			if (resolved_timestamp < 0)
-			{
-				// Even on the compatibility path, never allow the absolute
-				// timestamp to move below zero.
-				logae("Reject signed ext delta: sid=%u prev=%" PRId64 " raw=0x%08x signed=%" PRId32,
-					  stream_id,
-					  preceding_timestamp,
-					  timestamp_delta,
-					  signed_timestamp_delta);
-				return std::nullopt;
-			}
-
-			// This is the non-standard but tolerated case: a small negative
-			// extended delta that still resolves to a valid absolute timestamp.
-			logaw("Accept signed ext delta: sid=%u prev=%" PRId64 " raw=0x%08x signed=%" PRId32,
-				  stream_id,
-				  preceding_timestamp,
-				  timestamp_delta,
-				  signed_timestamp_delta);
-
-			return resolved_timestamp;
-		}
-	}
-
-	// Standard RTMP behavior: interpret the delta as an unsigned increment.
-	return preceding_timestamp + timestamp_delta;
-}
-
 RtmpChunkParser::ParseResult RtmpChunkParser::ParseMessageHeader(ov::ByteStream &stream, RtmpChunkHeader *chunk_header)
 {
 	auto &basic_header	 = chunk_header->basic_header;
@@ -430,8 +355,13 @@ RtmpChunkParser::ParseResult RtmpChunkParser::ParseMessageHeader(ov::ByteStream 
 
 			const auto parsed_timestamp		 = ParseTimestampField(
 				header.stream_id,
-				stream, chunk_header,
-				header.timestamp);
+				stream,
+				chunk_header->is_extended_timestamp,
+				chunk_header->extended_timestamp,
+				chunk_header->message_header_length,
+				header.timestamp,
+				EXTENDED_TIMESTAMP_INDICATOR,
+				EXTENDED_TIMESTAMP_SIZE);
 
 			if (parsed_timestamp.has_value() == false)
 
@@ -463,8 +393,13 @@ RtmpChunkParser::ParseResult RtmpChunkParser::ParseMessageHeader(ov::ByteStream 
 
 			const auto parsed_timestamp_delta = ParseTimestampField(
 				preceding_completed_header->stream_id,
-				stream, chunk_header,
-				header.timestamp_delta);
+				stream,
+				chunk_header->is_extended_timestamp,
+				chunk_header->extended_timestamp,
+				chunk_header->message_header_length,
+				header.timestamp_delta,
+				EXTENDED_TIMESTAMP_INDICATOR,
+				EXTENDED_TIMESTAMP_SIZE);
 
 			if (parsed_timestamp_delta.has_value() == false)
 			{
@@ -478,7 +413,9 @@ RtmpChunkParser::ParseResult RtmpChunkParser::ParseMessageHeader(ov::ByteStream 
 				preceding_completed_header->stream_id,
 				preceding_completed_header->timestamp,
 				parsed_timestamp_delta.value(),
-				chunk_header->is_extended_timestamp);
+				chunk_header->is_extended_timestamp,
+				SIGNED_EXTENDED_TIMESTAMP_DELTA_SIGN_BIT,
+				MAX_NEGATIVE_EXTENDED_TIMESTAMP_DELTA_MS);
 
 			if (resolved_timestamp.has_value() == false)
 			{
@@ -503,8 +440,13 @@ RtmpChunkParser::ParseResult RtmpChunkParser::ParseMessageHeader(ov::ByteStream 
 
 			const auto parsed_timestamp_delta = ParseTimestampField(
 				preceding_completed_header->stream_id,
-				stream, chunk_header,
-				header.timestamp_delta);
+				stream,
+				chunk_header->is_extended_timestamp,
+				chunk_header->extended_timestamp,
+				chunk_header->message_header_length,
+				header.timestamp_delta,
+				EXTENDED_TIMESTAMP_INDICATOR,
+				EXTENDED_TIMESTAMP_SIZE);
 
 			if (parsed_timestamp_delta.has_value() == false)
 			{
@@ -517,7 +459,9 @@ RtmpChunkParser::ParseResult RtmpChunkParser::ParseMessageHeader(ov::ByteStream 
 				preceding_completed_header->stream_id,
 				preceding_completed_header->timestamp,
 				parsed_timestamp_delta.value(),
-				chunk_header->is_extended_timestamp);
+				chunk_header->is_extended_timestamp,
+				SIGNED_EXTENDED_TIMESTAMP_DELTA_SIGN_BIT,
+				MAX_NEGATIVE_EXTENDED_TIMESTAMP_DELTA_MS);
 
 			if (resolved_timestamp.has_value() == false)
 			{
@@ -562,8 +506,13 @@ RtmpChunkParser::ParseResult RtmpChunkParser::ParseMessageHeader(ov::ByteStream 
 				{
 					const auto parsed_timestamp_field = ParseTimestampField(
 						completed.stream_id,
-						stream, chunk_header,
-						EXTENDED_TIMESTAMP_INDICATOR);
+						stream,
+						chunk_header->is_extended_timestamp,
+						chunk_header->extended_timestamp,
+						chunk_header->message_header_length,
+						EXTENDED_TIMESTAMP_INDICATOR,
+						EXTENDED_TIMESTAMP_INDICATOR,
+						EXTENDED_TIMESTAMP_SIZE);
 
 					if (parsed_timestamp_field.has_value() == false)
 					{
@@ -596,8 +545,13 @@ RtmpChunkParser::ParseResult RtmpChunkParser::ParseMessageHeader(ov::ByteStream 
 				// The origin message header is T0
 				const auto parsed_timestamp = ParseTimestampField(
 					completed.stream_id,
-					stream, chunk_header,
-					chunk_header->is_extended_timestamp ? EXTENDED_TIMESTAMP_INDICATOR : static_cast<uint32_t>(completed.timestamp));
+					stream,
+					chunk_header->is_extended_timestamp,
+					chunk_header->extended_timestamp,
+					chunk_header->message_header_length,
+					chunk_header->is_extended_timestamp ? EXTENDED_TIMESTAMP_INDICATOR : static_cast<uint32_t>(completed.timestamp),
+					EXTENDED_TIMESTAMP_INDICATOR,
+					EXTENDED_TIMESTAMP_SIZE);
 
 				if (parsed_timestamp.has_value() == false)
 				{
@@ -612,8 +566,13 @@ RtmpChunkParser::ParseResult RtmpChunkParser::ParseMessageHeader(ov::ByteStream 
 				// The origin message header is T1 or T2
 				const auto parsed_timestamp_delta = ParseTimestampField(
 					completed.stream_id,
-					stream, chunk_header,
-					chunk_header->is_extended_timestamp ? EXTENDED_TIMESTAMP_INDICATOR : completed.timestamp_delta);
+					stream,
+					chunk_header->is_extended_timestamp,
+					chunk_header->extended_timestamp,
+					chunk_header->message_header_length,
+					chunk_header->is_extended_timestamp ? EXTENDED_TIMESTAMP_INDICATOR : completed.timestamp_delta,
+					EXTENDED_TIMESTAMP_INDICATOR,
+					EXTENDED_TIMESTAMP_SIZE);
 
 				if (parsed_timestamp_delta.has_value() == false)
 				{
@@ -626,7 +585,9 @@ RtmpChunkParser::ParseResult RtmpChunkParser::ParseMessageHeader(ov::ByteStream 
 					completed.stream_id,
 					completed.timestamp,
 					parsed_timestamp_delta.value(),
-					chunk_header->is_extended_timestamp);
+					chunk_header->is_extended_timestamp,
+					SIGNED_EXTENDED_TIMESTAMP_DELTA_SIGN_BIT,
+					MAX_NEGATIVE_EXTENDED_TIMESTAMP_DELTA_MS);
 
 				if (resolved_timestamp.has_value() == false)
 				{
@@ -671,125 +632,6 @@ RtmpChunkParser::ParseResult RtmpChunkParser::ParseHeader(ov::ByteStream &stream
 	return ParseMessageHeader(stream, chunk_header);
 }
 
-int64_t RtmpChunkParser::CalculateRolledTimestamp(const uint32_t stream_id, const int64_t last_timestamp, int64_t parsed_timestamp)
-{
-	// RTMP timestamps are carried as 32-bit serial numbers on the wire.
-	// Keep the serial width, its modulo, and the RFC1982 half-range explicit so
-	// the rollover math below reads directly in terms of the specification.
-	constexpr int64_t SERIAL_BITS					  = 32;
-	constexpr int64_t SERIAL_MODULO					  = (1LL << SERIAL_BITS);
-	constexpr int64_t SERIAL_HALF_RANGE				  = (1LL << (SERIAL_BITS - 1));
-	// Some senders behave as if the timestamp space were signed int32_t and
-	// restart near INT32_MAX -> 0. Keep that compatibility boundary separate
-	// from the real RTMP modulo so the fallback path stays clearly scoped.
-	constexpr int64_t SIGNED_SERIAL_MODULO			  = (1LL << (SERIAL_BITS - 1));
-	// Only the low 32 bits participate in RFC1982 ordering; higher bits belong
-	// to older unfolded epochs reconstructed locally by the receiver.
-	constexpr uint64_t SERIAL_VALUE_MASK			  = 0xFFFFFFFFULL;
-	// Some senders appear to reset absolute timestamps at the signed 31-bit
-	// boundary instead of RTMP's unsigned 32-bit boundary. Treat that as a
-	// compatibility path only when the implied forward delta is very small.
-	constexpr int64_t MAX_SIGNED_WRAP_COMPAT_DELTA_MS = 10 * 1000;
-
-	// RFC1982 comparison works on the 32-bit serial value only, not on the full
-	// accumulated absolute timestamp. Masking with an unsigned 32-bit pattern
-	// intentionally strips any older epoch bits before converting to uint32_t.
-	const auto last_serial							  = static_cast<uint32_t>(last_timestamp & SERIAL_VALUE_MASK);
-	const auto parsed_serial						  = static_cast<uint32_t>(parsed_timestamp);
-
-	if (last_serial == parsed_serial)
-	{
-		return last_timestamp;
-	}
-
-	// RTMP specification
-	//
-	// Because timestamps are 32 bits long, they roll over every 49 days, 17
-	// hours, 2 minutes and 47.296 seconds. Because streams are allowed to
-	// run continuously, potentially for years on end, an RTMP application
-	// SHOULD use serial number arithmetic [RFC1982] when processing
-	// timestamps, and SHOULD be capable of handling wraparound. For
-	// example, an application assumes that all adjacent timestamps are
-	// within 2^31 - 1 milliseconds of each other, so 10000 comes after
-	// 4000000000, and 3000000000 comes before 4000000000.
-
-	// completed.timestamp calculated from this formula (https://tools.ietf.org/html/rfc1982#section-3.1):
-	//
-	// Serial numbers may be incremented by the addition of a positive
-	// integer n, where n is taken from the range of integers
-	// [0 .. (2^(SERIAL_BITS - 1) - 1)].  For a sequence number s, the
-	// result of such an addition, s', is defined as
-	//
-	//                 s' = (s + n) modulo (2 ^ SERIAL_BITS)
-	//
-	// where the addition and modulus operations here act upon values that
-	// are non-negative values of unbounded size in the usual ways of
-	// integer arithmetic.
-
-	// This helper is only for absolute timestamp paths (Type 0, or Type 3 that
-	// inherits Type 0 semantics). Type 1/2/3 delta paths already operate on an
-	// unfolded absolute timestamp plus a delta, so applying RFC1982 again there
-	// would be incorrect.
-	//
-	// Behavior:
-	// - if the parsed serial is after the previous serial and wrapped below it,
-	//   advance to the next 32-bit epoch
-	// - if the parsed serial is not after the previous serial, keep it in the
-	//   current/prior epoch so backward Type 0 timestamps remain backward
-	const int64_t serial_epoch_base = last_timestamp - static_cast<int64_t>(last_serial);
-	int64_t resolved_timestamp		= serial_epoch_base + parsed_serial;
-	const int64_t serial_diff		= static_cast<int64_t>(parsed_serial) - last_serial;
-	const bool parsed_serial_is_after_previous =
-		((serial_diff > 0) && (serial_diff < SERIAL_HALF_RANGE)) ||
-		((serial_diff < 0) && ((-serial_diff) > SERIAL_HALF_RANGE));
-
-	// Compatibility path for senders that appear to restart absolute
-	// timestamps at INT32_MAX -> 0 instead of UINT32_MAX -> 0. This is not
-	// standard RTMP rollover, so only accept it when the implied forward delta
-	// is very small.
-	const int64_t signed_wrap_forward_delta = SIGNED_SERIAL_MODULO - static_cast<int64_t>(last_serial) + parsed_serial;
-	const bool signed_wrap_compat_candidate =
-		(parsed_serial_is_after_previous == false) &&
-		(last_serial < SIGNED_SERIAL_MODULO) &&
-		(parsed_serial < SIGNED_SERIAL_MODULO) &&
-		(parsed_serial < last_serial) &&
-		(signed_wrap_forward_delta > 0) &&
-		(signed_wrap_forward_delta <= MAX_SIGNED_WRAP_COMPAT_DELTA_MS);
-
-	if (parsed_serial_is_after_previous)
-	{
-		if (resolved_timestamp <= last_timestamp)
-		{
-			resolved_timestamp += SERIAL_MODULO;
-			logad("Timestamp is rolled forward: last TS: %" PRId64 ", parsed: %" PRId64 ", resolved: %" PRId64,
-				  last_timestamp,
-				  parsed_timestamp,
-				  resolved_timestamp);
-		}
-	}
-	// Only reach this branch when the normal RFC1982 32-bit comparison says the
-	// parsed serial is not after the previous one.
-	else if (signed_wrap_compat_candidate)
-	{
-		resolved_timestamp = last_timestamp + signed_wrap_forward_delta;
-		logad("Timestamp is rolled forward with signed-31bit compatibility: last TS: %" PRId64 ", parsed: %" PRId64 ", delta: %" PRId64 ", resolved: %" PRId64,
-			  last_timestamp,
-			  parsed_timestamp,
-			  signed_wrap_forward_delta,
-			  resolved_timestamp);
-	}
-	else if (resolved_timestamp > last_timestamp)
-	{
-		resolved_timestamp -= SERIAL_MODULO;
-		logai("Timestamp is resolved as backward T0: last TS: %" PRId64 ", parsed: %" PRId64 ", resolved: %" PRId64,
-			  last_timestamp,
-			  parsed_timestamp,
-			  resolved_timestamp);
-	}
-
-	return resolved_timestamp;
-}
-
 std::shared_ptr<const RtmpMessage> RtmpChunkParser::GetMessage()
 {
 	if (_message_queue.IsEmpty())
@@ -810,20 +652,6 @@ std::shared_ptr<const RtmpMessage> RtmpChunkParser::GetMessage()
 size_t RtmpChunkParser::GetMessageCount() const
 {
 	return _message_queue.Size();
-}
-
-info::NamePath RtmpChunkParser::GetNamePath() const
-{
-	std::lock_guard lock_guard(_name_path_mutex);
-	return _name_path;
-}
-
-void RtmpChunkParser::UpdateNamePath(const info::NamePath &stream_name_path)
-{
-	std::lock_guard lock_guard(_name_path_mutex);
-
-	_name_path = stream_name_path;
-	_message_queue.SetAlias(ov::String::FormatString("RTMP queue for %s", _name_path.CStr()));
 }
 
 void RtmpChunkParser::Destroy()
