+++ src/projects/modules/rtmp/chunk/rtmp_chunk_parser.h
@@ -8,28 +8,16 @@
 //==============================================================================
 #pragma once
 
-#include <base/info/info.h>
-
-#include <deque>
-#include <map>
-#include <memory>
-#include <optional>
-
+#include "../rtmp_chunk_parser_common.h"
 #include "rtmp_datastructure.h"
 #include "rtmp_define.h"
 #include "rtmp_mux_util.h"
 
-class RtmpChunkParser
+class RtmpChunkParser : public RtmpChunkParserCommon<RtmpChunkHeader, RtmpMessage>
 {
 public:
-	enum class ParseResult
-	{
-		Error,
-		NeedMoreData,
-		Parsed,
-	};
+	using ParseResult = typename RtmpChunkParserCommon<RtmpChunkHeader, RtmpMessage>::ParseResult;
 
-public:
 	RtmpChunkParser(size_t chunk_size);
 	virtual ~RtmpChunkParser();
 
@@ -55,9 +43,6 @@ public:
 	{
 		_chunk_size = chunk_size;
 	}
-
-	info::NamePath GetNamePath() const;
-	void UpdateNamePath(const info::NamePath &stream_name_path);
 
 	void Destroy();
 
@@ -86,40 +71,6 @@ private:
 	/// @return `Parsed`, `NeedMoreData`, or `Error`.
 	ParseResult ParseBasicHeader(ov::ByteStream &stream, RtmpChunkHeader *chunk_header);
 
-	/// Parses a 24-bit timestamp field and its optional extended timestamp.
-	///
-	/// @param stream_id RTMP message stream identifier used for logging/context.
-	/// @param stream Input byte stream positioned at the timestamp field.
-	/// @param chunk_header Destination header object whose extended-timestamp
-	///        metadata is updated on success.
-	/// @param encoded_value The already-read 24-bit timestamp or timestamp-delta
-	///        field value.
-	///
-	/// @return The semantic timestamp value on success, or `std::nullopt` when
-	///         more bytes are required to finish parsing the field.
-	std::optional<uint32_t> ParseTimestampField(
-		const uint32_t stream_id,
-		ov::ByteStream &stream,
-		RtmpChunkHeader *chunk_header,
-		const uint32_t encoded_value);
-
-	/// Resolves a timestamp delta against the preceding absolute timestamp.
-	///
-	/// @param stream_id RTMP message stream identifier used for logging/context.
-	/// @param preceding_timestamp Previously resolved absolute timestamp.
-	/// @param timestamp_delta Parsed timestamp-delta field value.
-	/// @param is_extended_timestamp True when the delta came from the extended
-	///        timestamp field.
-	///
-	/// @return The resolved absolute timestamp, or `std::nullopt` when the
-	///         non-standard signed-delta compatibility path would produce an
-	///         invalid negative absolute timestamp.
-	std::optional<int64_t> ResolveTimestampDelta(
-		const uint32_t stream_id,
-		const int64_t preceding_timestamp,
-		const uint32_t timestamp_delta,
-		const bool is_extended_timestamp) const;
-
 	/// Parses the RTMP message header for an already-parsed basic header.
 	///
 	/// @param stream Input byte stream positioned at the RTMP message header.
@@ -135,31 +86,4 @@ private:
 	///
 	/// @return `Parsed`, `NeedMoreData`, or `Error`.
 	ParseResult ParseHeader(ov::ByteStream &stream, RtmpChunkHeader *chunk_header);
-
-	/// Unfolds a 32-bit absolute RTMP timestamp using RFC1982 serial arithmetic.
-	///
-	/// @param stream_id RTMP message stream identifier used for logging/context.
-	/// @param last_timestamp Previously resolved absolute timestamp.
-	/// @param parsed_timestamp Newly parsed 32-bit serial timestamp value.
-	///
-	/// @return The resolved absolute timestamp in the current or next serial
-	///         epoch.
-	int64_t CalculateRolledTimestamp(const uint32_t stream_id, const int64_t last_timestamp, int64_t parsed_timestamp);
-
-private:
-#if DEBUG
-	uint64_t _chunk_index	   = 0ULL;
-	uint64_t _total_read_bytes = 0ULL;
-#endif	// DEBUG
-
-	bool _need_to_parse_new_header = true;
-	std::shared_ptr<RtmpMessage> _current_message;
-	std::map<uint32_t, std::shared_ptr<RtmpMessage>> _pending_message_map;
-	std::map<uint32_t, std::shared_ptr<const RtmpChunkHeader>> _preceding_chunk_header_map;
-
-	ov::Queue<std::shared_ptr<const RtmpMessage>> _message_queue{nullptr, 500};
-	size_t _chunk_size;
-
-	mutable std::mutex _name_path_mutex;
-	info::NamePath _name_path;
 };
