+++ src/projects/modules/rtmp_v2/chunk/rtmp_chunk_parser.h
@@ -8,29 +8,17 @@
 //==============================================================================
 #pragma once
 
-#include <base/info/info.h>
-
-#include <deque>
-#include <map>
-#include <memory>
-#include <optional>
-
+#include "../../rtmp/rtmp_chunk_parser_common.h"
 #include "rtmp_datastructure.h"
 #include "rtmp_define.h"
 
 namespace modules::rtmp
 {
-	class ChunkParser
+	class ChunkParser : public ::RtmpChunkParserCommon<ChunkHeader, Message>
 	{
 	public:
-		enum class ParseResult
-		{
-			Error,
-			NeedMoreData,
-			Parsed,
-		};
+		using ParseResult = typename ::RtmpChunkParserCommon<ChunkHeader, Message>::ParseResult;
 
-	public:
 		ChunkParser(int chunk_size);
 		virtual ~ChunkParser();
 
@@ -51,8 +39,6 @@ namespace modules::rtmp
 
 		std::shared_ptr<const Message> GetMessage();
 		size_t GetMessageCount() const;
-		info::NamePath GetNamePath() const;
-		void UpdateNamePath(const info::NamePath &stream_name_path);
 
 		void SetChunkSize(size_t chunk_size)
 		{
@@ -92,40 +78,6 @@ namespace modules::rtmp
 		/// @return `Parsed`, `NeedMoreData`, or `Error`.
 		ParseResult ParseBasicHeader(ov::ByteStream &stream, ChunkHeader *chunk_header);
 
-		/// Parses a 24-bit timestamp field and its optional extended timestamp.
-		///
-		/// @param stream_id RTMP message stream identifier used for logging/context.
-		/// @param stream Input byte stream positioned at the timestamp field.
-		/// @param chunk_header Destination header object whose extended-timestamp
-		///        metadata is updated on success.
-		/// @param encoded_value The already-read 24-bit timestamp or timestamp-delta
-		///        field value.
-		///
-		/// @return The semantic timestamp value on success, or `std::nullopt` when
-		///         more bytes are required to finish parsing the field.
-		std::optional<uint32_t> ParseTimestampField(
-			const uint32_t stream_id,
-			ov::ByteStream &stream,
-			ChunkHeader *chunk_header,
-			const uint32_t encoded_value);
-
-		/// Resolves a timestamp delta against the preceding absolute timestamp.
-		///
-		/// @param stream_id RTMP message stream identifier used for logging/context.
-		/// @param preceding_timestamp Previously resolved absolute timestamp.
-		/// @param timestamp_delta Parsed timestamp-delta field value.
-		/// @param is_extended_timestamp True when the delta came from the extended
-		///        timestamp field.
-		///
-		/// @return The resolved absolute timestamp, or `std::nullopt` when the
-		///         non-standard signed-delta compatibility path would produce an
-		///         invalid negative absolute timestamp.
-		std::optional<int64_t> ResolveTimestampDelta(
-			const uint32_t stream_id,
-			const int64_t preceding_timestamp,
-			const uint32_t timestamp_delta,
-			const bool is_extended_timestamp) const;
-
 		/// Parses the RTMP message header for an already-parsed basic header.
 		///
 		/// @param stream Input byte stream positioned at the RTMP message header.
@@ -141,31 +93,5 @@ namespace modules::rtmp
 		///
 		/// @return `Parsed`, `NeedMoreData`, or `Error`.
 		ParseResult ParseHeader(ov::ByteStream &stream, ChunkHeader *chunk_header);
-
-		/// Unfolds a 32-bit absolute RTMP timestamp using RFC1982 serial arithmetic.
-		///
-		/// @param stream_id RTMP message stream identifier used for logging/context.
-		/// @param last_timestamp Previously resolved absolute timestamp.
-		/// @param parsed_timestamp Newly parsed 32-bit serial timestamp value.
-		///
-		/// @return The resolved absolute timestamp in the current or next serial
-		///         epoch.
-		int64_t CalculateRolledTimestamp(const uint32_t stream_id, const int64_t last_timestamp, int64_t parsed_timestamp);
-
-	private:
-#if DEBUG
-		uint64_t _chunk_index	   = 0ULL;
-		uint64_t _total_read_bytes = 0ULL;
-#endif	// DEBUG
-
-		bool _need_to_parse_new_header = true;
-		std::shared_ptr<Message> _current_message;
-		std::map<uint32_t, std::shared_ptr<Message>> _pending_message_map;
-		std::map<uint32_t, std::shared_ptr<const ChunkHeader>> _preceding_chunk_header_map;
-
-		ov::Queue<std::shared_ptr<const Message>> _message_queue{nullptr, 500};
-		mutable std::mutex _name_path_mutex;
-		info::NamePath _name_path;
-		size_t _chunk_size;
 	};
 }  // namespace modules::rtmp
