+++ src/projects/modules/rtmp_v2/rtmp_private.h
@@ -9,3 +9,6 @@
 #pragma once
 
 #define OV_LOG_TAG "Modules.RTMPv2"
+
+#define OV_LOG_PREFIX_FORMAT "[%s] "
+#define OV_LOG_PREFIX_VALUE GetNamePath().CStr()
